Ict showroom

+++ homeScreen/app/src/main/java/com/archery/tessa/homescreen/models/MeasuredDataSet.java
@@ -8,13 +8,13 @@ import java.util.List;
  */
 
 public class MeasuredDataSet {
-    private int timestamp;
+    private long timestamp;
     private int counter;
     private List<SensorData> sensors;
 
 
     public List<SensorData> getSensors(){return this.sensors;}
-    public int getTimestamp(){return this.timestamp;}
+    public long getTimestamp(){return this.timestamp;}
 
     public int getCounter() {
         return counter;
